+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 6.4.0
 - Raise an exception in submit! if the job serializes to a message that is
   above the native SQS limit for message size.
+- Ensure SendMessageBatch is only performed for batches totaling 256KB of message size or less.
 - Insert Sqewer::Error between StandardError and our custom errors for easier rescuing
 
 ### 6.3.0